+++ crates/evm/symbolic/src/runtime/solver.rs
@@ -1,5 +1,20 @@
 use super::*;
 
+mod hard_arith_fallback;
+mod monotonic_product;
+mod normalize;
+
+use hard_arith_fallback::constraints_prefer_hard_arith_fallback_first;
+pub(crate) use hard_arith_fallback::hard_arith_fallback_model;
+#[cfg(test)]
+pub(crate) use hard_arith_fallback::{expr_contains_hard_arith, fallback_single_var_model};
+#[cfg(test)]
+pub(crate) use monotonic_product::product_monotonic_unsat;
+use monotonic_product::product_monotonic_unsat_normalized;
+pub(crate) use normalize::normalize_constraints_for_solver;
+#[cfg(test)]
+pub(crate) use normalize::{normalize_bool_for_solver, normalize_expr_for_solver};
+
 /// Errors that arise when parsing or constructing solver commands from configuration.
 #[derive(Debug, thiserror::Error)]
 pub(crate) enum SolverConfigError {
@@ -235,14 +250,15 @@ impl SymbolicSolver for SmtLibSubprocessSolver {
         )
         .entered();
         trace!(query_id = self.queries, constraint_count = constraints.len(), "solver is_sat");
-        if product_monotonic_unsat(constraints) {
+        let smt_constraints = normalize_constraints_for_solver(constraints);
+        if product_monotonic_unsat_normalized(&smt_constraints) {
             trace!("is_sat: monotonic product contradiction");
             return Ok(false);
         }
-        let output = match self.query(constraints, false) {
+        let output = match self.query_normalized(&smt_constraints, false, constraints) {
             Ok(output) => output,
             Err(SymbolicError::SolverUnknown) => {
-                if hard_arith_fallback_model(constraints).is_some() {
+                if hard_arith_fallback_model(&smt_constraints).is_some() {
                     self.heuristic_witnesses += 1;
                     trace!("is_sat: hard arithmetic fallback model after solver unknown");
                     return Ok(true);
@@ -255,7 +271,7 @@ impl SymbolicSolver for SmtLibSubprocessSolver {
             "sat" => Ok(true),
             "unsat" => Ok(false),
             "unknown" => {
-                if hard_arith_fallback_model(constraints).is_some() {
+                if hard_arith_fallback_model(&smt_constraints).is_some() {
                     self.heuristic_witnesses += 1;
                     Ok(true)
                 } else {
@@ -278,15 +294,18 @@ impl SymbolicSolver for SmtLibSubprocessSolver {
         )
         .entered();
         trace!(query_id = self.queries, constraint_count = constraints.len(), "solver model");
-        if let Some(model) = hard_arith_fallback_model(constraints) {
+        let smt_constraints = normalize_constraints_for_solver(constraints);
+        if constraints_prefer_hard_arith_fallback_first(&smt_constraints)
+            && let Some(model) = hard_arith_fallback_model(&smt_constraints)
+        {
             self.heuristic_witnesses += 1;
-            trace!("model: hard arithmetic fallback model");
+            trace!("model: hard arithmetic fallback model before solver");
             return Ok(model);
         }
-        let output = match self.query(constraints, true) {
+        let output = match self.query_normalized(&smt_constraints, true, constraints) {
             Ok(output) => output,
             Err(SymbolicError::SolverUnknown) => {
-                if let Some(model) = hard_arith_fallback_model(constraints) {
+                if let Some(model) = hard_arith_fallback_model(&smt_constraints) {
                     self.heuristic_witnesses += 1;
                     trace!("model: hard arithmetic fallback model after solver unknown");
                     return Ok(model);
@@ -300,7 +319,7 @@ impl SymbolicSolver for SmtLibSubprocessSolver {
             "sat" => parse_and_validate_model(&output, constraints),
             "unsat" => Err(SymbolicError::Solver("counterexample path became unsat".to_string())),
             "unknown" => {
-                if let Some(model) = hard_arith_fallback_model(constraints) {
+                if let Some(model) = hard_arith_fallback_model(&smt_constraints) {
                     self.heuristic_witnesses += 1;
                     Ok(model)
                 } else {
@@ -347,15 +366,15 @@ impl SmtLibSubprocessSolver {
         }
     }
 
-    /// Implements the `query` solver helper.
-    pub(crate) fn query(
+    /// Sends already-normalized constraints to the configured solver portfolio.
+    pub(crate) fn query_normalized(
         &mut self,
-        constraints: &[BoolExpr],
+        smt_constraints: &[BoolExpr],
         model: bool,
+        model_constraints: &[BoolExpr],
     ) -> Result<String, SymbolicError> {
-        let smt_constraints = normalize_constraints_for_solver(constraints);
         let mut vars = BTreeSet::new();
-        for constraint in &smt_constraints {
+        for constraint in smt_constraints {
             constraint.collect_vars(&mut vars);
         }
 
@@ -374,7 +393,7 @@ impl SmtLibSubprocessSolver {
         for var in vars {
             let _ = writeln!(smt, "(declare-fun {var} () (_ BitVec 256))");
         }
-        for constraint in &smt_constraints {
+        for constraint in smt_constraints {
             let _ = writeln!(smt, "(assert {})", constraint.smt());
         }
         smt.push_str("(check-sat)\n");
@@ -387,7 +406,7 @@ impl SmtLibSubprocessSolver {
         }
 
         let result =
-            run_solver_commands(&commands, &smt, self.timeout, model.then_some(constraints));
+            run_solver_commands(&commands, &smt, self.timeout, model.then_some(model_constraints));
         self.portfolio_scheduler.record(&ordered_commands, &result.summaries);
         if self.dump_smt {
             self.portfolio_diagnostics.record(&result.summaries);
@@ -1299,1072 +1318,6 @@ pub(crate) fn validate_solver_model_output(
     constraints: &[BoolExpr],
 ) -> Result<(), SymbolicError> {
     parse_and_validate_model(output, constraints).map(|_| ())
-}
-
-/// Normalizes path constraints into an equivalent, solver-friendlier form.
-pub(crate) fn normalize_constraints_for_solver(constraints: &[BoolExpr]) -> Vec<BoolExpr> {
-    constraints.iter().cloned().map(normalize_bool_for_solver).collect()
-}
-
-/// Normalizes one boolean expression into an equivalent, solver-friendlier form.
-pub(crate) fn normalize_bool_for_solver(expr: BoolExpr) -> BoolExpr {
-    if let Some(normalized) = normalize_udiv_bool_for_solver(&expr) {
-        return normalized;
-    }
-
-    match expr {
-        BoolExpr::Const(value) => BoolExpr::Const(value),
-        BoolExpr::Not(value) => normalize_bool_for_solver(*value).not(),
-        BoolExpr::And(values) => {
-            BoolExpr::and(values.into_iter().map(normalize_bool_for_solver).collect())
-        }
-        BoolExpr::Eq(left, right) => {
-            let normalized =
-                BoolExpr::Eq(normalize_expr_for_solver(left), normalize_expr_for_solver(right));
-            normalize_udiv_bool_for_solver(&normalized).unwrap_or_else(|| match normalized {
-                BoolExpr::Eq(left, right) => BoolExpr::eq(left, right),
-                _ => unreachable!(),
-            })
-        }
-        BoolExpr::Cmp(op, left, right) => {
-            let normalized = BoolExpr::Cmp(
-                op,
-                normalize_expr_for_solver(left),
-                normalize_expr_for_solver(right),
-            );
-            normalize_udiv_bool_for_solver(&normalized).unwrap_or(normalized)
-        }
-    }
-}
-
-/// Normalizes one word expression into an equivalent, solver-friendlier form.
-pub(crate) fn normalize_expr_for_solver(expr: Expr) -> Expr {
-    if let Some(rebuilt) = rebuild_word_from_extracted_byte_terms(&expr)
-        && rebuilt != expr
-    {
-        return normalize_expr_for_solver(rebuilt);
-    }
-
-    match expr {
-        Expr::Const(_) | Expr::Var(_) | Expr::Keccak { .. } | Expr::Hash { .. } => expr,
-        Expr::Not(value) => Expr::Not(Box::new(normalize_expr_for_solver(*value))),
-        Expr::Op(op, left, right) => {
-            let left = normalize_expr_for_solver(*left);
-            let right = normalize_expr_for_solver(*right);
-            if matches!(op, ExprOp::Add | ExprOp::Mul | ExprOp::And | ExprOp::Or | ExprOp::Xor)
-                && right < left
-            {
-                Expr::op(op, right, left)
-            } else {
-                Expr::op(op, left, right)
-            }
-        }
-        Expr::Ite(cond, left, right) => normalize_ite_expr_for_solver(*cond, *left, *right),
-    }
-}
-
-/// Normalizes a word-valued conditional expression.
-pub(crate) fn normalize_ite_expr_for_solver(cond: BoolExpr, left: Expr, right: Expr) -> Expr {
-    let cond = normalize_bool_for_solver(cond);
-    let left = normalize_expr_for_solver(left);
-    let right = normalize_expr_for_solver(right);
-    if left == right {
-        return left;
-    }
-    if matches!(left, Expr::Const(value) if value == U256::from(1))
-        && bool_from_word_expr(&right).as_ref() == Some(&cond)
-    {
-        return right;
-    }
-    if matches!(right, Expr::Const(value) if value.is_zero())
-        && bool_from_word_expr(&left).as_ref() == Some(&cond)
-    {
-        return left;
-    }
-    Expr::Ite(Box::new(cond), Box::new(left), Box::new(right))
-}
-
-/// Rebuilds a word from OR-ed byte-extraction terms when the source is recoverable.
-pub(crate) fn rebuild_word_from_extracted_byte_terms(expr: &Expr) -> Option<Expr> {
-    let mut terms = Vec::new();
-    collect_or_terms(expr, &mut terms);
-    if terms.len() <= 1 {
-        return None;
-    }
-
-    let mut source = None;
-    let mut seen = [false; 32];
-    for term in terms {
-        if matches!(term, Expr::Const(value) if value.is_zero()) {
-            continue;
-        }
-        let (term_source, index) = extracted_shifted_byte_term(term)?;
-        match &source {
-            Some(source) if source != &term_source => return None,
-            Some(_) => {}
-            None => source = Some(term_source),
-        }
-        seen[index] = true;
-    }
-
-    let source = source?;
-    for (index, seen) in seen.into_iter().enumerate() {
-        if !seen && expr_known_byte(&source, index) != Some(0) {
-            return None;
-        }
-    }
-    Some(source)
-}
-
-/// Flattens nested bitwise-OR expressions into their leaf terms.
-pub(crate) fn collect_or_terms<'a>(expr: &'a Expr, terms: &mut Vec<&'a Expr>) {
-    match expr {
-        Expr::Op(ExprOp::Or, left, right) => {
-            collect_or_terms(left, terms);
-            collect_or_terms(right, terms);
-        }
-        expr => terms.push(expr),
-    }
-}
-
-/// Returns the source word and byte index for one shifted extracted-byte term.
-pub(crate) fn extracted_shifted_byte_term(term: &Expr) -> Option<(Expr, usize)> {
-    match term {
-        Expr::Op(ExprOp::Shl, byte, shift) => {
-            let Expr::Const(shift) = shift.as_ref() else { return None };
-            let shift = shift.to::<usize>();
-            if shift % 8 != 0 || shift > 248 {
-                return None;
-            }
-            let index = 31 - shift / 8;
-            let source = extracted_unshifted_byte_source(byte, index)?;
-            Some((source, index))
-        }
-        term => extracted_unshifted_byte_source(term, 31).map(|source| (source, 31)),
-    }
-}
-
-/// Returns the source word for an unshifted byte extraction at `index`.
-pub(crate) fn extracted_unshifted_byte_source(term: &Expr, index: usize) -> Option<Expr> {
-    let expr = strip_low_byte_mask(term)?;
-    if index == 31 {
-        return Some(expr.clone());
-    }
-    let Expr::Op(ExprOp::Shr, source, shift) = expr else { return None };
-    let Expr::Const(shift) = shift.as_ref() else { return None };
-    (*shift == U256::from((31 - index) * 8)).then(|| *source.clone())
-}
-
-/// Rewrites exact EVM unsigned-division zero/nonzero predicates without `bvudiv`.
-pub(crate) fn normalize_udiv_bool_for_solver(expr: &BoolExpr) -> Option<BoolExpr> {
-    match expr {
-        BoolExpr::Eq(left, Expr::Const(value)) if value.is_zero() => {
-            bool_from_word_expr(left).map(BoolExpr::not).or_else(|| {
-                if word_bool_always_true(left) {
-                    Some(BoolExpr::Const(false))
-                } else {
-                    normalize_udiv_eq_zero(left, &Expr::Const(U256::ZERO))
-                }
-            })
-        }
-        BoolExpr::Eq(Expr::Const(value), right) if value.is_zero() => {
-            bool_from_word_expr(right).map(BoolExpr::not).or_else(|| {
-                if word_bool_always_true(right) {
-                    Some(BoolExpr::Const(false))
-                } else {
-                    normalize_udiv_eq_zero(&Expr::Const(U256::ZERO), right)
-                }
-            })
-        }
-        BoolExpr::Eq(left, Expr::Const(value)) if *value == U256::from(1) => {
-            bool_from_word_expr(left)
-        }
-        BoolExpr::Eq(Expr::Const(value), right) if *value == U256::from(1) => {
-            bool_from_word_expr(right)
-        }
-        BoolExpr::Not(value) => match value.as_ref() {
-            BoolExpr::Cmp(op, left, right) => {
-                normalize_add_overflow_cmp_for_solver(*op, left, right)
-                    .map(BoolExpr::not)
-                    .or_else(|| normalize_udiv_cmp_for_solver(*op, left, right).map(BoolExpr::not))
-            }
-            BoolExpr::Eq(left, Expr::Const(value)) if value.is_zero() => {
-                if word_bool_always_true(left) {
-                    Some(BoolExpr::Const(true))
-                } else {
-                    normalize_udiv_eq_zero(left, &Expr::Const(U256::ZERO)).map(BoolExpr::not)
-                }
-            }
-            BoolExpr::Eq(Expr::Const(value), right) if value.is_zero() => {
-                if word_bool_always_true(right) {
-                    Some(BoolExpr::Const(true))
-                } else {
-                    normalize_udiv_eq_zero(&Expr::Const(U256::ZERO), right).map(BoolExpr::not)
-                }
-            }
-            BoolExpr::Eq(left, right) => normalize_udiv_eq_zero(left, right).map(BoolExpr::not),
-            _ => None,
-        },
-        BoolExpr::Eq(left, right) => normalize_udiv_eq_zero(left, right),
-        BoolExpr::Cmp(op, left, right) => normalize_add_overflow_cmp_for_solver(*op, left, right)
-            .or_else(|| normalize_udiv_cmp_for_solver(*op, left, right)),
-        BoolExpr::Const(_) | BoolExpr::And(_) => None,
-    }
-}
-
-/// Extracts the boolean condition represented by a word-valued `0`/`1` expression.
-pub(crate) fn bool_from_word_expr(expr: &Expr) -> Option<BoolExpr> {
-    let expr = strip_low_byte_mask(expr)?;
-    let Expr::Ite(condition, then_expr, else_expr) = expr else { return None };
-    match (then_expr.as_ref(), else_expr.as_ref()) {
-        (Expr::Const(then_value), Expr::Const(else_value))
-            if *then_value == U256::from(1) && else_value.is_zero() =>
-        {
-            Some(normalize_bool_for_solver((**condition).clone()))
-        }
-        (Expr::Const(then_value), Expr::Const(else_value))
-            if then_value.is_zero() && *else_value == U256::from(1) =>
-        {
-            Some(normalize_bool_for_solver((**condition).clone()).not())
-        }
-        _ => None,
-    }
-}
-
-/// Returns whether monotonic product facts make these constraints unsatisfiable.
-pub(crate) fn product_monotonic_unsat(constraints: &[BoolExpr]) -> bool {
-    let constraints = normalize_constraints_for_solver(constraints);
-    let mut less_than = BTreeSet::new();
-    let mut positive = BTreeSet::new();
-    for constraint in &constraints {
-        collect_order_facts(constraint, &mut less_than, &mut positive);
-    }
-
-    constraints.iter().any(|constraint| {
-        product_less_than_negation(constraint).is_some_and(|(left_a, left_b, right_a, right_b)| {
-            product_less_than_known(left_a, left_b, right_a, right_b, &less_than, &positive)
-        })
-    })
-}
-
-/// Collects simple unsigned ordering facts from normalized constraints.
-pub(crate) fn collect_order_facts(
-    expr: &BoolExpr,
-    less_than: &mut BTreeSet<(Expr, Expr)>,
-    positive: &mut BTreeSet<Expr>,
-) {
-    match expr {
-        BoolExpr::And(values) => {
-            for value in values {
-                collect_order_facts(value, less_than, positive);
-            }
-        }
-        BoolExpr::Cmp(BoolExprOp::Ult, left, right) => {
-            less_than.insert((left.clone(), right.clone()));
-            if matches!(left, Expr::Const(value) if value.is_zero()) {
-                positive.insert(right.clone());
-            }
-        }
-        BoolExpr::Cmp(BoolExprOp::Ugt, left, right) => {
-            less_than.insert((right.clone(), left.clone()));
-            if matches!(right, Expr::Const(value) if value.is_zero()) {
-                positive.insert(left.clone());
-            }
-        }
-        BoolExpr::Const(_) | BoolExpr::Not(_) | BoolExpr::Eq(_, _) | BoolExpr::Cmp(_, _, _) => {}
-    }
-}
-
-/// Extracts `!(a * b < c * d)` as product operands.
-pub(crate) fn product_less_than_negation(expr: &BoolExpr) -> Option<(&Expr, &Expr, &Expr, &Expr)> {
-    let BoolExpr::Not(value) = expr else { return None };
-    let BoolExpr::Cmp(BoolExprOp::Ult, left, right) = value.as_ref() else { return None };
-    let (left_a, left_b) = mul_operands(left)?;
-    let (right_a, right_b) = mul_operands(right)?;
-    Some((left_a, left_b, right_a, right_b))
-}
-
-/// Returns whether known facts imply `left_a * left_b < right_a * right_b`.
-pub(crate) fn product_less_than_known(
-    left_a: &Expr,
-    left_b: &Expr,
-    right_a: &Expr,
-    right_b: &Expr,
-    less_than: &BTreeSet<(Expr, Expr)>,
-    positive: &BTreeSet<Expr>,
-) -> bool {
-    product_less_than_known_ordered(left_a, left_b, right_a, right_b, less_than, positive)
-        || product_less_than_known_ordered(left_b, left_a, right_a, right_b, less_than, positive)
-        || product_less_than_known_ordered(left_a, left_b, right_b, right_a, less_than, positive)
-        || product_less_than_known_ordered(left_b, left_a, right_b, right_a, less_than, positive)
-}
-
-/// Checks the ordered monotonicity case `0 < a < c && 0 < b < d`.
-pub(crate) fn product_less_than_known_ordered(
-    left_a: &Expr,
-    left_b: &Expr,
-    right_a: &Expr,
-    right_b: &Expr,
-    less_than: &BTreeSet<(Expr, Expr)>,
-    positive: &BTreeSet<Expr>,
-) -> bool {
-    positive.contains(left_a)
-        && positive.contains(left_b)
-        && less_than.contains(&(left_a.clone(), right_a.clone()))
-        && less_than.contains(&(left_b.clone(), right_b.clone()))
-        && mul_cannot_overflow_256(left_a, left_b)
-        && mul_cannot_overflow_256(right_a, right_b)
-}
-
-/// Returns the operands for an unsigned multiplication expression.
-pub(crate) fn mul_operands(expr: &Expr) -> Option<(&Expr, &Expr)> {
-    match expr {
-        Expr::Op(ExprOp::Mul, left, right) => Some((left, right)),
-        _ => None,
-    }
-}
-
-/// Rewrites exact unsigned-addition overflow checks when operand bounds preclude overflow.
-pub(crate) fn normalize_add_overflow_cmp_for_solver(
-    op: BoolExprOp,
-    left: &Expr,
-    right: &Expr,
-) -> Option<BoolExpr> {
-    match op {
-        BoolExprOp::Ugt if add_overflow_check(left, right) => Some(BoolExpr::Const(false)),
-        BoolExprOp::Ult if add_overflow_check(right, left) => Some(BoolExpr::Const(false)),
-        _ => None,
-    }
-}
-
-/// Returns whether `left > left + increment` is an impossible overflow check.
-pub(crate) fn add_overflow_check(left: &Expr, right: &Expr) -> bool {
-    let Some((base, increment)) = add_with_operand(right, left) else { return false };
-    base == left && add_cannot_overflow_256(base, increment)
-}
-
-/// Returns the operands if `expr` is an addition involving `operand`.
-pub(crate) fn add_with_operand<'a>(expr: &'a Expr, operand: &Expr) -> Option<(&'a Expr, &'a Expr)> {
-    let Expr::Op(ExprOp::Add, left, right) = expr else { return None };
-    if left.as_ref() == operand {
-        Some((left, right))
-    } else if right.as_ref() == operand {
-        Some((right, left))
-    } else {
-        None
-    }
-}
-
-/// Returns whether unsigned addition of two expressions cannot overflow 256 bits.
-pub(crate) fn add_cannot_overflow_256(left: &Expr, right: &Expr) -> bool {
-    expr_unsigned_bits(left).max(expr_unsigned_bits(right)).saturating_add(1) <= 256
-}
-
-/// Returns whether a word-valued boolean expression is an exact tautology.
-pub(crate) fn word_bool_always_true(expr: &Expr) -> bool {
-    let mut terms = Vec::new();
-    collect_or_terms(expr, &mut terms);
-    if terms.len() <= 1 {
-        return false;
-    }
-
-    let bool_terms = terms.iter().filter_map(|term| word_bool_term(term)).collect::<Vec<_>>();
-    for zero_term in &bool_terms {
-        let Some(zero_operand) = zero_check_operand(zero_term) else { continue };
-        if bool_terms.iter().any(|term| checked_mul_guard_for_operand(term, zero_operand)) {
-            return true;
-        }
-    }
-    false
-}
-
-/// Converts one `0`/`1` word boolean term into its boolean condition.
-pub(crate) fn word_bool_term(expr: &Expr) -> Option<&BoolExpr> {
-    let Expr::Ite(condition, then_expr, else_expr) = expr else { return None };
-    match (then_expr.as_ref(), else_expr.as_ref()) {
-        (Expr::Const(then_value), Expr::Const(else_value))
-            if *then_value == U256::from(1) && else_value.is_zero() =>
-        {
-            Some(condition)
-        }
-        _ => None,
-    }
-}
-
-/// Returns the operand tested by `operand == 0`.
-pub(crate) fn zero_check_operand(expr: &BoolExpr) -> Option<&Expr> {
-    match expr {
-        BoolExpr::Eq(left, Expr::Const(value)) if value.is_zero() => Some(left),
-        BoolExpr::Eq(Expr::Const(value), right) if value.is_zero() => Some(right),
-        _ => None,
-    }
-}
-
-/// Returns whether this condition is Solidity's checked-mul guard for `zero_operand`.
-pub(crate) fn checked_mul_guard_for_operand(expr: &BoolExpr, zero_operand: &Expr) -> bool {
-    let BoolExpr::Eq(left, right) = expr else { return false };
-    checked_mul_guard_side(left, right, zero_operand)
-        || checked_mul_guard_side(right, left, zero_operand)
-}
-
-/// Matches `a == 0 ? 0 : (a * b) / a` against the expected quotient.
-pub(crate) fn checked_mul_guard_side(
-    div_expr: &Expr,
-    expected: &Expr,
-    zero_operand: &Expr,
-) -> bool {
-    let Expr::Ite(condition, then_expr, else_expr) = div_expr else { return false };
-    if zero_check_operand(condition).is_none_or(|operand| operand != zero_operand) {
-        return false;
-    }
-    if !matches!(then_expr.as_ref(), Expr::Const(value) if value.is_zero()) {
-        return false;
-    }
-    let Some((numerator, denominator)) = udiv_operands(else_expr) else { return false };
-    if denominator != zero_operand {
-        return false;
-    }
-    let Expr::Op(ExprOp::Mul, left, right) = numerator else { return false };
-    let other = if left.as_ref() == zero_operand {
-        right.as_ref()
-    } else if right.as_ref() == zero_operand {
-        left.as_ref()
-    } else {
-        return false;
-    };
-    other == expected && mul_cannot_overflow_256(zero_operand, other)
-}
-
-/// Returns whether unsigned multiplication of two expressions cannot overflow 256 bits.
-pub(crate) fn mul_cannot_overflow_256(left: &Expr, right: &Expr) -> bool {
-    expr_unsigned_bits(left).saturating_add(expr_unsigned_bits(right)) <= 256
-}
-
-/// Returns a conservative unsigned bit-width upper bound for an expression.
-pub(crate) fn expr_unsigned_bits(expr: &Expr) -> usize {
-    match expr {
-        Expr::Const(value) => value.bit_len().max(1),
-        Expr::Op(ExprOp::And, left, right) => match (left.as_ref(), right.as_ref()) {
-            (expr, Expr::Const(mask)) | (Expr::Const(mask), expr) => {
-                expr_unsigned_bits(expr).min(mask.bit_len())
-            }
-            _ => 256,
-        },
-        Expr::Op(ExprOp::Add, left, right) => {
-            expr_unsigned_bits(left).max(expr_unsigned_bits(right)).saturating_add(1).min(256)
-        }
-        Expr::Op(ExprOp::Mul, left, right) => {
-            expr_unsigned_bits(left).saturating_add(expr_unsigned_bits(right)).min(256)
-        }
-        Expr::Op(ExprOp::UDiv, left, _) => expr_unsigned_bits(left),
-        Expr::Ite(_, left, right) => expr_unsigned_bits(left).max(expr_unsigned_bits(right)),
-        _ => 256,
-    }
-}
-
-/// Rewrites `udiv(a, b) == 0` predicates using EVM division-by-zero semantics.
-pub(crate) fn normalize_udiv_eq_zero(left: &Expr, right: &Expr) -> Option<BoolExpr> {
-    if matches!(right, Expr::Const(value) if value.is_zero())
-        && let Some(condition) = normalize_expr_eq_zero_for_solver(left)
-    {
-        return Some(condition);
-    }
-    if matches!(left, Expr::Const(value) if value.is_zero())
-        && let Some(condition) = normalize_expr_eq_zero_for_solver(right)
-    {
-        return Some(condition);
-    }
-    None
-}
-
-/// Rewrites `expr == 0` when `expr` contains exactly-normalizable unsigned division.
-pub(crate) fn normalize_expr_eq_zero_for_solver(expr: &Expr) -> Option<BoolExpr> {
-    if let Some((numerator, denominator)) = udiv_operands(expr) {
-        return Some(udiv_zero_condition(numerator, denominator));
-    }
-    if let Expr::Ite(condition, then_expr, else_expr) = expr {
-        let then_zero = normalize_expr_eq_zero_for_solver(then_expr).unwrap_or_else(|| {
-            BoolExpr::eq(normalize_expr_for_solver((**then_expr).clone()), Expr::Const(U256::ZERO))
-        });
-        let else_zero = normalize_expr_eq_zero_for_solver(else_expr).unwrap_or_else(|| {
-            BoolExpr::eq(normalize_expr_for_solver((**else_expr).clone()), Expr::Const(U256::ZERO))
-        });
-        if then_zero.smt().contains("bvudiv") || else_zero.smt().contains("bvudiv") {
-            return None;
-        }
-        return Some(BoolExpr::or(vec![
-            BoolExpr::and(vec![normalize_bool_for_solver((**condition).clone()), then_zero]),
-            BoolExpr::and(vec![normalize_bool_for_solver((**condition).clone()).not(), else_zero]),
-        ]));
-    }
-    None
-}
-
-/// Rewrites `expr != 0` when `expr` contains exactly-normalizable unsigned division.
-pub(crate) fn normalize_expr_ne_zero_for_solver(expr: &Expr) -> Option<BoolExpr> {
-    if let Some((numerator, denominator)) = udiv_operands(expr) {
-        return Some(udiv_nonzero_condition(numerator, denominator));
-    }
-    if let Expr::Ite(condition, then_expr, else_expr) = expr {
-        let then_nonzero = normalize_expr_ne_zero_for_solver(then_expr).unwrap_or_else(|| {
-            BoolExpr::eq(normalize_expr_for_solver((**then_expr).clone()), Expr::Const(U256::ZERO))
-                .not()
-        });
-        let else_nonzero = normalize_expr_ne_zero_for_solver(else_expr).unwrap_or_else(|| {
-            BoolExpr::eq(normalize_expr_for_solver((**else_expr).clone()), Expr::Const(U256::ZERO))
-                .not()
-        });
-        if then_nonzero.smt().contains("bvudiv") || else_nonzero.smt().contains("bvudiv") {
-            return None;
-        }
-        return Some(BoolExpr::or(vec![
-            BoolExpr::and(vec![normalize_bool_for_solver((**condition).clone()), then_nonzero]),
-            BoolExpr::and(vec![
-                normalize_bool_for_solver((**condition).clone()).not(),
-                else_nonzero,
-            ]),
-        ]));
-    }
-    None
-}
-
-/// Rewrites `udiv(a, b)` zero/nonzero comparisons using EVM division-by-zero semantics.
-pub(crate) fn normalize_udiv_cmp_for_solver(
-    op: BoolExprOp,
-    left: &Expr,
-    right: &Expr,
-) -> Option<BoolExpr> {
-    match (op, left, right) {
-        (BoolExprOp::Ugt, div, Expr::Const(value)) if value.is_zero() => {
-            normalize_expr_ne_zero_for_solver(div)
-        }
-        (BoolExprOp::Uge, div, Expr::Const(value)) if *value == U256::from(1) => {
-            normalize_expr_ne_zero_for_solver(div)
-        }
-        (BoolExprOp::Ule, div, Expr::Const(value)) if value.is_zero() => {
-            normalize_expr_eq_zero_for_solver(div)
-        }
-        (BoolExprOp::Ult, div, Expr::Const(value)) if *value == U256::from(1) => {
-            normalize_expr_eq_zero_for_solver(div)
-        }
-        (BoolExprOp::Ult, Expr::Const(value), div) if value.is_zero() => {
-            normalize_expr_ne_zero_for_solver(div)
-        }
-        (BoolExprOp::Ule, Expr::Const(value), div) if *value == U256::from(1) => {
-            normalize_expr_ne_zero_for_solver(div)
-        }
-        (BoolExprOp::Uge, Expr::Const(value), div) if value.is_zero() => {
-            normalize_expr_eq_zero_for_solver(div)
-        }
-        (BoolExprOp::Ugt, Expr::Const(value), div) if *value == U256::from(1) => {
-            normalize_expr_eq_zero_for_solver(div)
-        }
-        _ => None,
-    }
-}
-
-/// Returns the operands for an unsigned division expression.
-pub(crate) fn udiv_operands(expr: &Expr) -> Option<(&Expr, &Expr)> {
-    match expr {
-        Expr::Op(ExprOp::UDiv, numerator, denominator) => Some((numerator, denominator)),
-        _ => None,
-    }
-}
-
-/// Builds the exact condition for EVM `udiv(numerator, denominator) == 0`.
-pub(crate) fn udiv_zero_condition(numerator: &Expr, denominator: &Expr) -> BoolExpr {
-    BoolExpr::or(vec![
-        BoolExpr::eq(normalize_expr_for_solver(denominator.clone()), Expr::Const(U256::ZERO)),
-        BoolExpr::cmp(
-            BoolExprOp::Ult,
-            normalize_expr_for_solver(numerator.clone()),
-            normalize_expr_for_solver(denominator.clone()),
-        ),
-    ])
-}
-
-/// Builds the exact condition for EVM `udiv(numerator, denominator) != 0`.
-pub(crate) fn udiv_nonzero_condition(numerator: &Expr, denominator: &Expr) -> BoolExpr {
-    BoolExpr::and(vec![
-        BoolExpr::eq(normalize_expr_for_solver(denominator.clone()), Expr::Const(U256::ZERO)).not(),
-        BoolExpr::cmp(
-            BoolExprOp::Uge,
-            normalize_expr_for_solver(numerator.clone()),
-            normalize_expr_for_solver(denominator.clone()),
-        ),
-    ])
-}
-
-/// Returns the `bool_contains_hard_arith` solver helper result.
-pub(crate) fn bool_contains_hard_arith(expr: &BoolExpr) -> bool {
-    match expr {
-        BoolExpr::Const(_) => false,
-        BoolExpr::Not(value) => bool_contains_hard_arith(value),
-        BoolExpr::And(values) => values.iter().any(bool_contains_hard_arith),
-        BoolExpr::Eq(left, right) | BoolExpr::Cmp(_, left, right) => {
-            expr_contains_hard_arith(left) || expr_contains_hard_arith(right)
-        }
-    }
-}
-
-/// Returns the `expr_contains_hard_arith` solver helper result.
-pub(crate) fn expr_contains_hard_arith(expr: &Expr) -> bool {
-    match expr {
-        Expr::Const(_) | Expr::Var(_) | Expr::Keccak { .. } | Expr::Hash { .. } => false,
-        Expr::Not(value) => expr_contains_hard_arith(value),
-        Expr::Op(ExprOp::Mul, left, right) => expr_contains_var(left) && expr_contains_var(right),
-        Expr::Op(ExprOp::UDiv | ExprOp::URem | ExprOp::SDiv | ExprOp::SRem, left, right) => {
-            expr_contains_var(left) || expr_contains_var(right)
-        }
-        Expr::Op(_, left, right) => {
-            expr_contains_hard_arith(left) || expr_contains_hard_arith(right)
-        }
-        Expr::Ite(cond, left, right) => {
-            bool_contains_hard_arith(cond)
-                || expr_contains_hard_arith(left)
-                || expr_contains_hard_arith(right)
-        }
-    }
-}
-
-/// Returns whether the expression contains symbolic hash variables that local search should avoid.
-pub(crate) fn expr_contains_symbolic_hash(expr: &Expr) -> bool {
-    match expr {
-        Expr::Hash { .. } => true,
-        Expr::Keccak { len, bytes, .. } => {
-            expr_contains_symbolic_hash(len) || bytes.iter().any(expr_contains_symbolic_hash)
-        }
-        Expr::Const(_) | Expr::Var(_) => false,
-        Expr::Not(value) => expr_contains_symbolic_hash(value),
-        Expr::Op(_, left, right) => {
-            expr_contains_symbolic_hash(left) || expr_contains_symbolic_hash(right)
-        }
-        Expr::Ite(cond, left, right) => {
-            bool_contains_symbolic_hash(cond)
-                || expr_contains_symbolic_hash(left)
-                || expr_contains_symbolic_hash(right)
-        }
-    }
-}
-
-/// Returns whether the boolean expression contains symbolic hash variables.
-pub(crate) fn bool_contains_symbolic_hash(expr: &BoolExpr) -> bool {
-    match expr {
-        BoolExpr::Const(_) => false,
-        BoolExpr::Not(value) => bool_contains_symbolic_hash(value),
-        BoolExpr::And(values) => values.iter().any(bool_contains_symbolic_hash),
-        BoolExpr::Eq(left, right) | BoolExpr::Cmp(_, left, right) => {
-            expr_contains_symbolic_hash(left) || expr_contains_symbolic_hash(right)
-        }
-    }
-}
-
-/// Returns the `expr_contains_var` solver helper result.
-pub(crate) fn expr_contains_var(expr: &Expr) -> bool {
-    match expr {
-        Expr::Const(_) => false,
-        Expr::Var(_) | Expr::Keccak { .. } | Expr::Hash { .. } => true,
-        Expr::Not(value) => expr_contains_var(value),
-        Expr::Op(_, left, right) => expr_contains_var(left) || expr_contains_var(right),
-        Expr::Ite(cond, left, right) => {
-            bool_contains_var(cond) || expr_contains_var(left) || expr_contains_var(right)
-        }
-    }
-}
-
-/// Returns the `bool_contains_var` solver helper result.
-pub(crate) fn bool_contains_var(expr: &BoolExpr) -> bool {
-    match expr {
-        BoolExpr::Const(_) => false,
-        BoolExpr::Not(value) => bool_contains_var(value),
-        BoolExpr::And(values) => values.iter().any(bool_contains_var),
-        BoolExpr::Eq(left, right) | BoolExpr::Cmp(_, left, right) => {
-            expr_contains_var(left) || expr_contains_var(right)
-        }
-    }
-}
-
-/// Implements the `hard_arith_fallback_model` solver helper.
-pub(crate) fn hard_arith_fallback_model(
-    constraints: &[BoolExpr],
-) -> Option<BTreeMap<String, U256>> {
-    if !constraints.iter().any(bool_contains_hard_arith)
-        || constraints.iter().any(bool_contains_symbolic_hash)
-    {
-        return None;
-    }
-
-    let mut vars = BTreeSet::new();
-    let mut constants = BTreeSet::new();
-    for constraint in constraints {
-        collect_bool_fallback_vars(constraint, &mut vars);
-        collect_bool_constants(constraint, &mut constants);
-    }
-    let vars = fallback_search_vars(vars);
-    if vars.is_empty() || vars.len() > HARD_ARITH_FALLBACK_MAX_VARS {
-        return None;
-    }
-
-    let candidates = vars
-        .iter()
-        .map(|var| fallback_candidates_for_var(var, constraints, &constants))
-        .collect::<Option<Vec<_>>>()?;
-    let searched_vars = vars.iter().cloned().collect::<BTreeSet<_>>();
-    let constraint_vars = constraints
-        .iter()
-        .map(|constraint| {
-            let mut vars = BTreeSet::new();
-            constraint.collect_vars(&mut vars);
-            vars
-        })
-        .collect::<Vec<_>>();
-    let mut model = BTreeMap::new();
-    let mut assignments = 0usize;
-    let search = FallbackSearch {
-        constraints,
-        constraint_vars: &constraint_vars,
-        searched_vars: &searched_vars,
-        vars: &vars,
-        candidates: &candidates,
-    };
-    search.model(0, &mut model, &mut assignments)
-}
-
-/// Selects direct symbolic inputs for bounded fallback search.
-pub(crate) fn fallback_search_vars(vars: BTreeSet<String>) -> Vec<String> {
-    if vars.len() <= HARD_ARITH_FALLBACK_MAX_VARS {
-        return vars.into_iter().collect();
-    }
-
-    vars.into_iter()
-        .filter(|var| {
-            var.starts_with("calldata")
-                || var.starts_with("sequence")
-                || var.starts_with("create_address")
-                || var.starts_with("create2_address")
-                || !var.contains('_')
-        })
-        .collect()
-}
-
-/// Returns deterministic local-search candidates for one symbolic variable.
-pub(crate) fn fallback_candidates_for_var(
-    var: &str,
-    constraints: &[BoolExpr],
-    constants: &BTreeSet<U256>,
-) -> Option<Vec<U256>> {
-    let hints = MaskHints::for_var(var, constraints);
-    if (hints.one & hints.zero) != U256::ZERO {
-        return None;
-    }
-
-    let mut candidates = BTreeSet::new();
-    for candidate in [
-        U256::ZERO,
-        U256::from(1),
-        U256::from(2),
-        U256::from(3),
-        U256::MAX,
-        U256::MAX - U256::from(1),
-        U256::MAX - U256::from(2),
-    ] {
-        push_fallback_candidate(&mut candidates, candidate, hints);
-    }
-
-    for constant in constants.iter().copied() {
-        push_fallback_candidate(&mut candidates, constant, hints);
-        push_fallback_candidate(&mut candidates, constant.wrapping_add(U256::from(1)), hints);
-        push_fallback_candidate(&mut candidates, constant.wrapping_sub(U256::from(1)), hints);
-        if candidates.len() >= HARD_ARITH_FALLBACK_MAX_CANDIDATES_PER_VAR {
-            break;
-        }
-    }
-
-    for bit in 0..256 {
-        let power = U256::from(1) << bit;
-        push_fallback_candidate(&mut candidates, power, hints);
-        if candidates.len() >= HARD_ARITH_FALLBACK_MAX_CANDIDATES_PER_VAR {
-            break;
-        }
-    }
-
-    Some(candidates.into_iter().take(HARD_ARITH_FALLBACK_MAX_CANDIDATES_PER_VAR).collect())
-}
-
-/// Holds immutable state for recursive hard-arithmetic fallback search.
-struct FallbackSearch<'a> {
-    constraints: &'a [BoolExpr],
-    constraint_vars: &'a [BTreeSet<String>],
-    searched_vars: &'a BTreeSet<String>,
-    vars: &'a [String],
-    candidates: &'a [Vec<U256>],
-}
-
-impl FallbackSearch<'_> {
-    /// Searches the bounded candidate product for a satisfying assignment.
-    fn model(
-        &self,
-        index: usize,
-        model: &mut BTreeMap<String, U256>,
-        assignments: &mut usize,
-    ) -> Option<BTreeMap<String, U256>> {
-        if index == self.vars.len() {
-            *assignments += 1;
-            if *assignments > HARD_ARITH_FALLBACK_MAX_ASSIGNMENTS {
-                return None;
-            }
-            return fallback_model_satisfies_all_constraints(self.constraints, model)
-                .then(|| model.clone());
-        }
-
-        for candidate in &self.candidates[index] {
-            model.insert(self.vars[index].clone(), *candidate);
-            if fallback_partial_model_satisfies_known_constraints(
-                self.constraints,
-                self.constraint_vars,
-                self.searched_vars,
-                model,
-            ) && let Some(model) = self.model(index + 1, model, assignments)
-            {
-                return Some(model);
-            }
-            if *assignments > HARD_ARITH_FALLBACK_MAX_ASSIGNMENTS {
-                return None;
-            }
-        }
-        model.remove(&self.vars[index]);
-        None
-    }
-}
-
-/// Checks all constraints before returning a hard-arithmetic fallback witness.
-pub(crate) fn fallback_model_satisfies_all_constraints(
-    constraints: &[BoolExpr],
-    model: &BTreeMap<String, U256>,
-) -> bool {
-    constraints.iter().all(|constraint| eval_bool_expr(constraint, model).unwrap_or(false))
-}
-
-/// Checks constraints that depend only on already-assigned fallback variables.
-pub(crate) fn fallback_partial_model_satisfies_known_constraints(
-    constraints: &[BoolExpr],
-    constraint_vars: &[BTreeSet<String>],
-    searched_vars: &BTreeSet<String>,
-    model: &BTreeMap<String, U256>,
-) -> bool {
-    constraints.iter().zip(constraint_vars).all(|(constraint, vars)| {
-        !vars.is_subset(searched_vars)
-            || !vars.iter().all(|var| model.contains_key(var))
-            || eval_bool_expr(constraint, model).unwrap_or(false)
-    })
-}
-
-/// Collects variables that local hard-arithmetic search can assign directly.
-pub(crate) fn collect_bool_fallback_vars(expr: &BoolExpr, vars: &mut BTreeSet<String>) {
-    match expr {
-        BoolExpr::Const(_) => {}
-        BoolExpr::Not(value) => collect_bool_fallback_vars(value, vars),
-        BoolExpr::And(values) => {
-            for value in values {
-                collect_bool_fallback_vars(value, vars);
-            }
-        }
-        BoolExpr::Eq(left, right) | BoolExpr::Cmp(_, left, right) => {
-            collect_expr_fallback_vars(left, vars);
-            collect_expr_fallback_vars(right, vars);
-        }
-    }
-}
-
-/// Collects assignable variables from an expression, recursing into recomputable hashes.
-pub(crate) fn collect_expr_fallback_vars(expr: &Expr, vars: &mut BTreeSet<String>) {
-    match expr {
-        Expr::Const(_) | Expr::Hash { .. } => {}
-        Expr::Var(var) => {
-            vars.insert(var.clone());
-        }
-        Expr::Keccak { len, bytes, .. } => {
-            collect_expr_fallback_vars(len, vars);
-            for byte in bytes {
-                collect_expr_fallback_vars(byte, vars);
-            }
-        }
-        Expr::Not(value) => collect_expr_fallback_vars(value, vars),
-        Expr::Op(_, left, right) => {
-            collect_expr_fallback_vars(left, vars);
-            collect_expr_fallback_vars(right, vars);
-        }
-        Expr::Ite(cond, left, right) => {
-            collect_bool_fallback_vars(cond, vars);
-            collect_expr_fallback_vars(left, vars);
-            collect_expr_fallback_vars(right, vars);
-        }
-    }
-}
-
-/// Implements the `fallback_single_var_model` solver helper.
-#[cfg(test)]
-pub(crate) fn fallback_single_var_model(
-    constraints: &[BoolExpr],
-) -> Option<BTreeMap<String, U256>> {
-    let mut vars = BTreeSet::new();
-    let mut constants = BTreeSet::new();
-    for constraint in constraints {
-        constraint.collect_vars(&mut vars);
-        collect_bool_constants(constraint, &mut constants);
-    }
-
-    let var = if vars.len() == 1 { vars.iter().next()?.clone() } else { return None };
-    let hints = MaskHints::for_var(&var, constraints);
-    if (hints.one & hints.zero) != U256::ZERO {
-        return None;
-    }
-
-    let mut candidates = BTreeSet::new();
-    for candidate in [
-        U256::ZERO,
-        U256::from(1),
-        U256::from(2),
-        U256::MAX,
-        U256::MAX - U256::from(1),
-        U256::MAX - U256::from(2),
-    ] {
-        push_fallback_candidate(&mut candidates, candidate, hints);
-    }
-
-    for constant in constants.iter().copied() {
-        push_fallback_candidate(&mut candidates, constant, hints);
-        push_fallback_candidate(&mut candidates, constant.wrapping_add(U256::from(1)), hints);
-        push_fallback_candidate(&mut candidates, constant.wrapping_sub(U256::from(1)), hints);
-    }
-
-    for bit in 0..256 {
-        let power = U256::from(1) << bit;
-        push_fallback_candidate(&mut candidates, power, hints);
-        for constant in constants.iter().copied().take(64) {
-            push_fallback_candidate(&mut candidates, power | constant, hints);
-            push_fallback_candidate(&mut candidates, power.wrapping_add(constant), hints);
-        }
-    }
-
-    for candidate in candidates {
-        let model = BTreeMap::from([(var.clone(), candidate)]);
-        if constraints.iter().all(|constraint| eval_bool_expr(constraint, &model).unwrap_or(false))
-        {
-            return Some(model);
-        }
-    }
-
-    None
-}
-
-/// Applies the `push_fallback_candidate` solver helper.
-pub(crate) fn push_fallback_candidate(
-    candidates: &mut BTreeSet<U256>,
-    candidate: U256,
-    hints: MaskHints,
-) {
-    candidates.insert((candidate | hints.one) & !hints.zero);
-}
-
-/// Implements the `collect_bool_constants` solver helper.
-pub(crate) fn collect_bool_constants(expr: &BoolExpr, constants: &mut BTreeSet<U256>) {
-    match expr {
-        BoolExpr::Const(_) => {}
-        BoolExpr::Not(value) => collect_bool_constants(value, constants),
-        BoolExpr::And(values) => {
-            for value in values {
-                collect_bool_constants(value, constants);
-            }
-        }
-        BoolExpr::Eq(left, right) | BoolExpr::Cmp(_, left, right) => {
-            collect_expr_constants(left, constants);
-            collect_expr_constants(right, constants);
-        }
-    }
-}
-
-/// Implements the `collect_expr_constants` solver helper.
-pub(crate) fn collect_expr_constants(expr: &Expr, constants: &mut BTreeSet<U256>) {
-    match expr {
-        Expr::Const(value) => {
-            constants.insert(*value);
-        }
-        Expr::Var(_) | Expr::Keccak { .. } | Expr::Hash { .. } => {}
-        Expr::Not(value) => collect_expr_constants(value, constants),
-        Expr::Op(_, left, right) => {
-            collect_expr_constants(left, constants);
-            collect_expr_constants(right, constants);
-        }
-        Expr::Ite(cond, left, right) => {
-            collect_bool_constants(cond, constants);
-            collect_expr_constants(left, constants);
-            collect_expr_constants(right, constants);
-        }
-    }
-}
-
-#[derive(Clone, Copy, Debug, Default)]
-pub(crate) struct MaskHints {
-    pub(crate) one: U256,
-    pub(crate) zero: U256,
-}
-
-impl MaskHints {
-    /// Implements the `for_var` solver helper.
-    pub(crate) fn for_var(var: &str, constraints: &[BoolExpr]) -> Self {
-        let mut hints = Self::default();
-        for constraint in constraints {
-            hints.apply_bool(var, constraint, false);
-        }
-        hints
-    }
-
-    /// Applies the `apply_bool` solver helper.
-    pub(crate) fn apply_bool(&mut self, var: &str, expr: &BoolExpr, inverted: bool) {
-        match expr {
-            BoolExpr::Const(_) => {}
-            BoolExpr::Not(value) => self.apply_bool(var, value, !inverted),
-            BoolExpr::And(values) if !inverted => {
-                for value in values {
-                    self.apply_bool(var, value, false);
-                }
-            }
-            BoolExpr::Eq(left, right) => self.apply_equality(var, left, right, inverted),
-            BoolExpr::Cmp(_, _, _) | BoolExpr::And(_) => {}
-        }
-    }
-
-    /// Applies the `apply_equality` solver helper.
-    pub(crate) fn apply_equality(&mut self, var: &str, left: &Expr, right: &Expr, inverted: bool) {
-        if let Some(mask) =
-            zero_mask_equality(var, left, right).or_else(|| zero_mask_equality(var, right, left))
-        {
-            if inverted {
-                self.one |= mask;
-            } else {
-                self.zero |= mask;
-            }
-        }
-    }
-}
-
-/// Implements the `zero_mask_equality` solver helper.
-pub(crate) fn zero_mask_equality(var: &str, masked: &Expr, zero: &Expr) -> Option<U256> {
-    if !matches!(zero, Expr::Const(value) if value.is_zero()) {
-        return None;
-    }
-    match masked {
-        Expr::Op(ExprOp::And, left, right) => match (left.as_ref(), right.as_ref()) {
-            (Expr::Var(name), Expr::Const(mask)) | (Expr::Const(mask), Expr::Var(name))
-                if name == var =>
-            {
-                Some(*mask)
-            }
-            _ => None,
-        },
-        _ => None,
-    }
 }
 
 /// Returns the `parse_model` solver helper result.
